+++ drivers/bitplane/palette.c
@@ -39,14 +39,14 @@ long CDECL c_get_colour(Virtual *vwk, long colour)
 }
 
 
-void CDECL c_get_colours(Virtual *vwk, long colour, short *foreground, short* background)
+void CDECL c_get_colours(Virtual *vwk, long colour, long *foreground, long *background)
 {
   *foreground = x_get_colour(vwk->real_address, colour & 0xffff);
   *background = x_get_colour(vwk->real_address, (colour >> 16) & 0xffff);
 }
 
 
-void CDECL x_get_colours(Workstation *wk, long colour, short *foreground, short* background)
+void CDECL x_get_colours(Workstation *wk, long colour, short *foreground, short *background)
 {
   *foreground = x_get_colour(wk, colour & 0xffff);
   *background = x_get_colour(wk, (colour >> 16) & 0xffff);
